Added structure to keep api endpoints seperate
