+++ eagleEvents/templates/seating-chart.html.j2
@@ -136,7 +136,7 @@
     }
 
 
-    fetch('/changeSeat', {
+    fetch('/api/event/changeSeat', {
         credentials: "same-origin",
         headers: {
             'Content-Type': "application/json",
